Add iterator over files with invalid checksums (#17)
Previously, the only means to verify the on-disk file hashes via this library threw an exception on the first invalid file. This makes it possible to first check the hash list's authenticity with `Verifier::verifyCsigMessage()`, then construct a `ChecksumList` iterator with it, and finally provide the `ChecksumList` to a `FailedCheckumFilter` in order to list all invalid files.